Refactored stringifier
