+++ src/main/index.ts
@@ -1,12 +1,2 @@
-import {parseJson} from './parseJson';
-import {createJsonStringifier} from './createJsonStringifier';
-
-export * from './createJsonStringifier';
-export * from './types';
-
-export const parse = parseJson;
-
-export const stringify = createJsonStringifier({
-  isBigInt: (value) => typeof value === 'bigint',
-  stringifyBigInt: String,
-});
+export {parseJson as parse} from './parseJson';
+export {stringifyJson as stringify, StringifyJsonOptions} from './stringifyJson';
