+++ src/main/utils.ts
@@ -1,3 +1,3 @@
-export function die(message?: string, offset?: number): never {
-  throw new SyntaxError(message == null || offset == null ? message : message + ' at position ' + offset);
+export function die(message: string, offset: number): never {
+  throw new SyntaxError(message + ' at position ' + offset);
 }
