+++ src/test/stringifyJson.test.ts
@@ -1,12 +1,7 @@
-import {createJsonStringifier} from '../main';
+import {stringifyJson} from '../main/stringifyJson';
 import {parseJson} from '../main/parseJson';
 
-describe('createJsonStringifier', () => {
-
-  const stringifyJson = createJsonStringifier({
-    isBigInt: (value) => typeof value === 'bigint',
-    stringifyBigInt: String,
-  });
+describe('stringifyJson', () => {
 
   test('stringifies bigint', () => {
     expect(stringifyJson({foo: BigInt(123)})).toBe('{"foo":123}');
